+++ nfm-world/mad/interp/Interpolation.cs
@@ -7,18 +7,18 @@ public class Interpolation {
         return interp;
     }
 
-    public static int InterpolateAngle(int current, int prev, float ratio) {
-        int diff = current - prev;
+    public static float InterpolateAngle(float current, float prev, float ratio) {
+        float diff = current - prev;
         /*
          * this could go 359->1 or 1->359, so check the size of the movement, if too big
          * then subtract
          */
-        if (diff > 270) {
-            diff -= 360;
-        } else if (diff < -270) {
-            diff += 360;
+        if (diff > 270f) {
+            diff -= 360f;
+        } else if (diff < -270f) {
+            diff += 360f;
         }
-        int interp = (int)((diff * ratio) + prev);
+        float interp = (diff * ratio) + prev;
         return interp;
     }
 }
